pyproject.toml
[tool.poetry]
name = "scoolarcbackend"
version = "0.1.0"
description = "School Management System"
authors = ["abdullahzulfiqar653 <abdullahzulfiqar653@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.16"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
shortuuid = "^1.0.13"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
whitenoise = "^6.7.0"
gunicorn = "^23.0.0"
ipython = "^8.27.0"
django-soft-delete = "^1.0.14"
pillow = "^10.4.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
